fix: use add to hs
